image69.jpg
This is image is of Darjeeling Monastery, During the Chinese occupation of 1959 in Tibet many high ranking abbots fled to India and took refuge in the monastery. In 1961, Dhardo Rimpoche became head of the Yiga Choeling monastery Ghoom, Darjeeling. He died in 1990 and three years later, a boy named Tenzin Legshad Wangdi was recognised as his reincarnation.